+++ extensions/iceworks-codemod/src/logger/index.ts
@@ -1,0 +1,6 @@
+import getLogger from '@iceworks/logger';
+
+// eslint-disable-next-line
+const { name } = require('../../package.json');
+
+export default getLogger(name);