+++ extensions/iceworks-project-creator/web/src/pages/CreateProject/index.tsx
@@ -118,7 +118,9 @@ const CreateProject: React.FC = () => {
   async function onOpenFolderDialog() {
     try {
       const projectPath = await callService('project', 'getFolderPath');
-      setCurProjectField({ ...curProjectField, projectPath });
+      if (projectPath) {
+        setCurProjectField({ ...curProjectField, projectPath });
+      }
     } catch (e) {
       // ignore
     }