+++ extensions/iceworks-refactor/src/refactor/index.ts
@@ -1,0 +1,2 @@
+export { default as removeComponent } from './removeComponent';
+export { default as removeUnreferencedCode } from './removeUnreferencedCode';