+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.4.12 
+
+- fix: save user config data in template preview.
+- fix: select project path undefined
+
 ## 0.4.11
 
 - fix: AliInternal auto login